If the server closes (very likely), disconnect immediately
Servers like to minimise the amount of state they maintain so they
will close their side of the TCP connection. For example Google's DNS
will do this after being idle for about 1s. There's no point in keeping
the request sending side open because although `write` will succeed
in the sense that data will be transmitted, no response is possible.

When we know that the server has closed and no response is possible,
we should call `disconnect`. Note this will fail any requests which
have just been transmitted-- we still need to consider retransmitting
these in a future patch.

Signed-off-by: David Scott <dave@recoil.org>