Rewrite DNS query ids
When we receive queries over multiple TCP connections, beware that
the received query ids are scoped to the individual connections. For
example, 10 clients may connect to us and all submit a request with
transaction id 1. Since we are multiplexing requests onto a single
outgoing connection we must generate fresh ids for each outgoing
request and remember the mapping from generated fresh ids to received
ids.

Signed-off-by: David Scott <dave@recoil.org>