test: add a query counter and verify requests go only to the right server
Previously this test case would verify the result of a DNS lookup,
however it is possible to get the right answer for the wrong reason:
if the request was sent to both servers then the response from the
"right" server might arrive first.

This patch adds query counters to the server and verifies -- once the
system threads have idled -- the counter for the right server is 1 and
the counter for the wrong server is 0.

Fixes #10

Signed-off-by: David Scott <dave@recoil.org>