Remove getclientname
The original intention was to record the source IP of the outgoing
packets in the message recording callback, however we don't actually
know the source IP used by the kernel. Typically `getsockname` will
retur `0.0.0.0`. This patch removes this facility and makes the
addresses optional in the callback.

See feedback on [docker/vpnkit#132]

Signed-off-by: David Scott <dave@recoil.org>