Send requests to servers in groups
Previously we always had a flat list of servers to query. This patch
changes the signature of `choose_server` to return a list of lists of
servers and the request is sent to each group in series. Since
`choose_server` returns at most one group, no change can be observed.

Signed-off-by: David Scott <dave@recoil.org>