Rpc: handle races between disconnect and receive
If a response is late then it might arrive just before or just after
a disconnect. Since the wakeners hashtable entry is owned by the
client thread, it's possible for the `Lwt.wakeup_later` to be called
a second time, raising `Invalid_argument`.

For the dispatcher this was probably harmless since that thread is
about to fail in `Packet.read` on the next time round the loop. It's
still worth catching and logging because it represents a response
that was slightly too late, which could mean the timeouts are wrong.

For `disconnect` it would fail to clean up the rest of the threads.
Higher-level code would probably call `Lwt.cancel` but it's good not
to rely on this.

Signed-off-by: David Scott <dave.scott@docker.com>