If a single server fails, don't cancel all other attempts
Previously we would try to send a query to a set of servers
concurrently. The first successful response would be returned; but
the first failure would cause all other requests to be cancelled.
This is obviously wrong if you consider the case where a server
address is unreachable and the `sendto` fails quickly.

This patch gives every thread a chance to discover a positive
result and will only return a failure if the timeout hits.

Signed-off-by: David Scott <dave@recoil.org>