Fix problem where request ids could be re-used over a timeout
If a timeout occurs then the id becomes re-usable, and then we
could transmit a request and immediately receive the old reply.

This patch fixes the issue by storing the expected question along
with the id and dropping any responses which are sent to the
correct id but to the wrong question.

Signed-off-by: David Scott <dave@recoil.org>