resolver: query the cache per-server
Previously we treated the cache as a global map of question to answers.
This patch treats the cache as a per-server map of question to answers,
so that we always send queries to other servers even in the case of
one entry from another server being cached.

This is part of the decoupling of caching (of server responses) from the
logic of deciding which response to give the client.

Signed-off-by: David Scott <dave@recoil.org>