resolver: don't cancel RPCs when a response arrives
We wish to cache all the responses, even if they arrive a little
later. It seems polite to wait for all responses that we ask for,
even if we decide not to use them. When the responses are in
the cache, we won't ask the server again until after the TTL.

Signed-off-by: David Scott <dave@recoil.org>