Don't exclude servers which have a zone configured
Previously we would match the domain in the request against the
zone and

1. if there was any match, we would send the request only to matching
   servers
2. if there was no match, we would send the request only to servers
   with no zone

Rule (1) is good because it prevents the accidental leakage of private
DNS names so we keep it.

Rule (2) is bad because a server can be needed for general queries:
for example consider the situation when there is only one DNS server
configured which happens to have a zone.

This patch changes rule (2): now if there is no match, we will send the
(public) request to all configured servers.

Signed-off-by: David Scott <dave@recoil.org>