resolver: avoid blocking if a bad server is early on the servers list
Previously we would send all the queries concurrently but then used
`Lwt_list.fold_left_s` to walk over this list to look for the best response.
Although we would short-circuit the walk once we found a good response,
if a good response has arrived but happens to be on the tail of the list
we would wait for every other response before we find it. This could lead to
timeouts if a bad server is nearer the beginning of the list than a
good server.

This patch uses `nchoose_split` to process the results in thread-completion
order -- rather than arbitrary list order -- ensuring that we never block
when potentially good responses have arrived.

Signed-off-by: David Scott <dave@recoil.org>