resolver: wait for responses in equal priority groups
Previously we would take the first received successful response
which means that if we have a low priority server which returns
a different result, we would take that without waiting for the
high-priority server.

This patch divides the servers into equal priority groups and waits
for the first response within each group, and then each group one at
a time.

Signed-off-by: David Scott <dave@recoil.org>