resolver: send requests to all servers, only wait for replies from online servers
Previously we would send requests to all servers and wait for replies from
all servers, irrespective of whether they are online or not. Since
we wait for the timeout within each priority group, an offline server in
the first group would delay anything.

This patch still sends requests to all relevant servers -- this is important
to test if a server has come back online -- but will only wait for replies
from known-online servers. This means that when a server first goes offline
we will still wait for a timeout from it, but subsequently we will skip it.
When it comes back online we will drop the first response but respect the
later ones.

Signed-off-by: David Scott <dave@recoil.org>