Add a "should be cached" notion to a connection
The TCP protocol is quite expensive and we would like to cache open
connections. The UDP protocol is very cheap and we would prefer to
use fresh local port numbers to increase the entropy in every request.

This patch adds a notion of a connection that "should be cached" or
not, with Tcp being "true" and Udp being "false".

Signed-off-by: David Scott <dave@recoil.org>