test: remove hardcoded port numbers
Previously we hardcoded a set of port numbers which made it difficult
to add tests. The `find_free_address` returned a port which is
currently free, but didn't allocate it -- this made it difficult to use
e.g.

  let a1 = find_free_address () in
  let a2 = find_free_address () in
  assert (a1 <> a2) (* oops, they are the same *)

The new function `fresh_port` returns a fresh port number which is
guaranteed to be unique.

Signed-off-by: David Scott <dave@recoil.org>